Add 0x0c page breaks between existing solved chapters
Those page break markers display something like "^L" in both
vim and emacs.

In emacs, we can insert them with `C-q C-l` and navigate between them
with `C-x [` and `C-x ]`.

In vim, the digraph is `FF`, which means we can insert them with `C-k
FF`, which can also be used for searching and thus navigating between
those page breaks.